// src/features/ProductList/model/enums.ts
export enum ProductListStages {
    LOADING_PRODUCTS = 'LOADING_PRODUCTS',
    PICKING_PRODUCTS = 'PICKING_PRODUCTS',
}

export enum ProductListTransitionTypes {
    // Inner
    LOADING_PRODUCTS__PICKING_PRODUCTS = 'LOADING_PRODUCTS__PICKING_PRODUCTS',
    PICKING_PRODUCTS__LOADING_PRODUCTS = 'PICKING_PRODUCTS__LOADING_PRODUCTS',
    PICKING_PRODUCTS__PICKING_PRODUCTS = 'PICKING_PRODUCTS__PICKING_PRODUCTS',
    // To product info
    PICKING_PRODUCTS__PRODUCT_INFO = 'PICKING_PRODUCTS__PRODUCT_INFO',
}